Se agrega neurona Sigmoid.

+++ perceptron01/src/tests/TestLogicGates.java
@@ -28,93 +28,93 @@ public class TestLogicGates {
 
   @Test
   public void testNandGate() {
-    List<Integer> input = new ArrayList<Integer>(2);
-    input.add(0, 0);
-    input.add(1, 0);
+    List<Double> input = new ArrayList<Double>(2);
+    input.add(0, 0.0);
+    input.add(1, 0.0);
     nandNetwork.simulate(input);
-    assertEquals(new Integer(1), nandNetwork.calculateResult().get(0));
-    input.set(0, 0);
-    input.set(1, 1);
+    assertEquals(new Double(1), nandNetwork.calculateResult().get(0));
+    input.set(0, 0.0);
+    input.set(1, 1.0);
     nandNetwork.simulate(input);
-    assertEquals(new Integer(1), nandNetwork.calculateResult().get(0));
-    input.set(0, 1);
-    input.set(1, 0);
+    assertEquals(new Double(1), nandNetwork.calculateResult().get(0));
+    input.set(0, 1.0);
+    input.set(1, 0.0);
     nandNetwork.simulate(input);
-    assertEquals(new Integer(1), nandNetwork.calculateResult().get(0));
-    input.set(0, 1);
-    input.set(1, 1);
+    assertEquals(new Double(1), nandNetwork.calculateResult().get(0));
+    input.set(0, 1.0);
+    input.set(1, 1.0);
     nandNetwork.simulate(input);
-    assertEquals(new Integer(0), nandNetwork.calculateResult().get(0));
+    assertEquals(new Double(0), nandNetwork.calculateResult().get(0));
   }
 
   @Test
   public void testAndGate() {
-    List<Integer> input = new ArrayList<Integer>(2);
-    input.add(0, 0);
-    input.add(1, 0);
+    List<Double> input = new ArrayList<Double>(2);
+    input.add(0, 0.0);
+    input.add(1, 0.0);
     andNetwork.simulate(input);
-    assertEquals(new Integer(0), andNetwork.calculateResult().get(0));
-    input.set(0, 0);
-    input.set(1, 1);
+    assertEquals(new Double(0), andNetwork.calculateResult().get(0));
+    input.set(0, 0.0);
+    input.set(1, 1.0);
     andNetwork.simulate(input);
-    assertEquals(new Integer(0), andNetwork.calculateResult().get(0));
-    input.set(0, 1);
-    input.set(1, 0);
+    assertEquals(new Double(0), andNetwork.calculateResult().get(0));
+    input.set(0, 1.0);
+    input.set(1, 0.0);
     andNetwork.simulate(input);
-    assertEquals(new Integer(0), andNetwork.calculateResult().get(0));
-    input.set(0, 1);
-    input.set(1, 1);
+    assertEquals(new Double(0), andNetwork.calculateResult().get(0));
+    input.set(0, 1.0);
+    input.set(1, 1.0);
     andNetwork.simulate(input);
-    assertEquals(new Integer(1), andNetwork.calculateResult().get(0));
+    assertEquals(new Double(1), andNetwork.calculateResult().get(0));
   }
 
   @Test
   public void testOrGate() {
-    List<Integer> input = new ArrayList<Integer>(2);
-    input.add(0, 0);
-    input.add(1, 0);
+    List<Double> input = new ArrayList<Double>(2);
+    input.add(0, 0.0);
+    input.add(1, 0.0);
     orNetwork.simulate(input);
-    assertEquals(new Integer(0), orNetwork.calculateResult().get(0));
-    input.set(0, 0);
-    input.set(1, 1);
+    assertEquals(new Double(0), orNetwork.calculateResult().get(0));
+    input.set(0, 0.0);
+    input.set(1, 1.0);
     orNetwork.simulate(input);
-    assertEquals(new Integer(1), orNetwork.calculateResult().get(0));
-    input.set(0, 1);
-    input.set(1, 0);
+    assertEquals(new Double(1), orNetwork.calculateResult().get(0));
+    input.set(0, 1.0);
+    input.set(1, 0.0);
     orNetwork.simulate(input);
-    assertEquals(new Integer(1), orNetwork.calculateResult().get(0));
-    input.set(0, 1);
-    input.set(1, 1);
+    assertEquals(new Double(1), orNetwork.calculateResult().get(0));
+    input.set(0, 1.0);
+    input.set(1, 1.0);
     orNetwork.simulate(input);
-    assertEquals(new Integer(1), orNetwork.calculateResult().get(0));
+    assertEquals(new Double(1), orNetwork.calculateResult().get(0));
   }
 
   @Test
   public void testSumGate() {
-    List<Integer> input = new ArrayList<Integer>(2);
-    List<Integer> expected = new ArrayList<Integer>(2);
-    input.add(0, 0);
-    input.add(1, 0);
-    expected.add(0, 0);
-    expected.add(1, 0);
+    List<Double> input = new ArrayList<Double>(2);
+    List<Double> expected = new ArrayList<Double>(2);
+    input.add(0, 0.0);
+    input.add(1, 0.0);
+    expected.add(0, 0.0);
+    expected.add(1, 0.0);
     sumNetwork.simulate(input);
     assertEquals(expected, sumNetwork.calculateResult());
-    input.set(0, 1);
-    input.set(1, 0);
-    expected.set(0, 1);
-    expected.set(1, 0);
+    input.set(0, 1.0);
+    input.set(1, 0.0);
+    expected.set(0, 1.0);
+    expected.set(1, 0.0);
     sumNetwork.simulate(input);
     assertEquals(expected, sumNetwork.calculateResult());
-    input.set(0, 0);
-    input.set(1, 1);
-    expected.set(0, 1);
-    expected.set(1, 0);
+    input.set(0, 0.0);
+    input.set(1, 1.0);
+    expected.set(0, 1.0);
+    expected.set(1, 0.0);
     sumNetwork.simulate(input);
     assertEquals(expected, sumNetwork.calculateResult());
-    input.set(0, 1);
-    input.set(1, 1);
-    expected.set(0, 0);
-    expected.set(1, 1);
+    input.set(0, 1.0);
+    input.set(1, 1.0);
+    expected.set(0, 0.0);
+    expected.set(1, 1.0);
     sumNetwork.simulate(input);
     assertEquals(expected, sumNetwork.calculateResult());
   }
